Move api/halls/ to api/admins/
