+++ src/Api/admins/halls/api.js
@@ -1,4 +1,4 @@
-import axios from '../../config/axios';
+import axios from '../../../config/axios';
 
 export const fetchHalls = async () => {
   let halls = [];
