+++ src/Pages/Hall/update.jsx
@@ -1,7 +1,7 @@
 import Swal from 'sweetalert2';
 import { useState } from 'react';
 import { FaEdit } from 'react-icons/fa';
-import { updateHall } from '../../Api/halls/api';
+import { updateHall } from '../../Api/admins/halls/api';
 import Modal from '../../components/Modal';
 
 const Update = ({ data, setChanged }) => {
